+++ echopype/clean/__init__.py
@@ -1,6 +1,9 @@
-from .api import estimate_noise, remove_noise
+from .api import estimate_noise, remove_noise, get_impulse_noise_mask, get_transient_noise_mask
+
 
 __all__ = [
     "estimate_noise",
     "remove_noise",
+    "get_impulse_noise_mask",
+    "get_transient_noise_mask",
 ]
