+++ echopype/mask/__init__.py
@@ -1,10 +1,7 @@
 from .api import apply_mask, frequency_differencing
-from .mask_impulse_noise import get_impulse_noise_mask
-from .mask_transient_noise import get_transient_noise_mask
+
 
 __all__ = [
     "frequency_differencing",
     "apply_mask",
-    "get_transient_noise_mask",
-    "get_impulse_noise_mask",
 ]
